+++ lebedev.alexandr/P4/main.cpp
@@ -1,5 +1,5 @@
 #include <iostream>
-#include "string.hpp"
+#include "myOwnString.hpp"
 
 int main()
 {
