+++ lebedev.alexandr/P4/myOwnString.hpp
